pre_save signal tail decide to skip updates
When an update doesn't change a field that would sync, there's no
sense in doing a sync, and it can even result in a user not being able
to log into a service if the sync fails for some reason and last_login
changes (which we don't sync).

So provide a pre_save signal for dealing with that case.
If we're given update_fields, we use that to check if any of them are
fields that would sync. Otherwise we compare the updated object
against the one currently in the database and determine if any of the
changed fields are sync fields.

Tests are provided for all of the above.

Signed-off-by: Rick Elrod <rick@elrod.me>